Revert "Update code admin/users"
This reverts commit 51a9e973a1673bb71ee8cb21a0b62164ee41bb06.
