+++ src/frontend/src/components/sideNav/SideNav.tsx
@@ -12,7 +12,6 @@ import { NavItems } from "@/config";
 import { cn } from "@/lib/utils";
 import { ChevronLeft, ChevronRight } from "lucide-react";
 import { useSession } from "next-auth/react";
-import { SideNavItem } from "./SideNavItem";
 
 export default function SideNav() {
   const [isSidebarExpanded, setIsSidebarExpanded] = useState(true);
@@ -37,7 +36,7 @@ export default function SideNav() {
     setIsSidebarExpanded(!isSidebarExpanded);
   };
 
-  const navItems = NavItems(session?.user.role || "");
+  const navItems = NavItems(session?.user.role || '');
 
   return (
     <div suppressHydrationWarning>
@@ -110,59 +109,59 @@ export default function SideNav() {
   );
 }
 
-// export const SideNavItem: React.FC<{
-//   label: string;
-//   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-//   icon: any;
-//   path: string;
-//   active: boolean;
-//   isSidebarExpanded: boolean;
-// }> = ({ label, icon, path, active, isSidebarExpanded }) => {
-//   return (
-//     <>
-//       {isSidebarExpanded ? (
-//         <Link
-//           href={path}
-//           className={`h-full relative flex items-center whitespace-nowrap rounded-md ${
-//             active
-//               ? "font-base text-sm bg-neutral-200 shadow-sm text-neutral-700 dark:bg-neutral-800 dark:text-white"
-//               : "hover:bg-neutral-200 hover:text-neutral-700 dark:text-neutral-400 dark:hover:bg-neutral-800 dark:hover:text-white"
-//           }`}
-//           suppressHydrationWarning
-//         >
-//           <div className="relative font-base text-sm py-1.5 px-2 flex flex-row items-center space-x-2 rounded-md duration-100">
-//             {icon}
-//             <span>{label}</span>
-//           </div>
-//         </Link>
-//       ) : (
-//         <TooltipProvider delayDuration={70}>
-//           <Tooltip>
-//             <TooltipTrigger>
-//               <Link
-//                 href={path}
-//                 className={`h-full relative flex items-center whitespace-nowrap rounded-md ${
-//                   active
-//                     ? "font-base text-sm bg-neutral-200 text-neutral-700 dark:bg-neutral-800 dark:text-white"
-//                     : "hover:bg-neutral-200 hover:text-neutral-700 text-neutral-500 dark:text-neutral-400 dark:hover:bg-neutral-800 dark:hover:text-white"
-//                 }`}
-//                 suppressHydrationWarning={true}
-//               >
-//                 <div className="relative font-base text-sm p-2 flex flex-row items-center space-x-2 rounded-md duration-100">
-//                   {icon}
-//                 </div>
-//               </Link>
-//             </TooltipTrigger>
-//             <TooltipContent
-//               side="left"
-//               className="px-3 py-1.5 text-xs"
-//               sideOffset={10}
-//             >
-//               <span>{label}</span>
-//             </TooltipContent>
-//           </Tooltip>
-//         </TooltipProvider>
-//       )}
-//     </>
-//   );
-// };
+export const SideNavItem: React.FC<{
+  label: string;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  icon: any;
+  path: string;
+  active: boolean;
+  isSidebarExpanded: boolean;
+}> = ({ label, icon, path, active, isSidebarExpanded }) => {
+  return (
+    <>
+      {isSidebarExpanded ? (
+        <Link
+          href={path}
+          className={`h-full relative flex items-center whitespace-nowrap rounded-md ${
+            active
+              ? "font-base text-sm bg-neutral-200 shadow-sm text-neutral-700 dark:bg-neutral-800 dark:text-white"
+              : "hover:bg-neutral-200 hover:text-neutral-700 dark:text-neutral-400 dark:hover:bg-neutral-800 dark:hover:text-white"
+          }`}
+          suppressHydrationWarning
+        >
+          <div className="relative font-base text-sm py-1.5 px-2 flex flex-row items-center space-x-2 rounded-md duration-100">
+            {icon}
+            <span>{label}</span>
+          </div>
+        </Link>
+      ) : (
+        <TooltipProvider delayDuration={70}>
+          <Tooltip>
+            <TooltipTrigger>
+              <Link
+                href={path}
+                className={`h-full relative flex items-center whitespace-nowrap rounded-md ${
+                  active
+                    ? "font-base text-sm bg-neutral-200 text-neutral-700 dark:bg-neutral-800 dark:text-white"
+                    : "hover:bg-neutral-200 hover:text-neutral-700 text-neutral-500 dark:text-neutral-400 dark:hover:bg-neutral-800 dark:hover:text-white"
+                }`}
+                suppressHydrationWarning={true}
+              >
+                <div className="relative font-base text-sm p-2 flex flex-row items-center space-x-2 rounded-md duration-100">
+                  {icon}
+                </div>
+              </Link>
+            </TooltipTrigger>
+            <TooltipContent
+              side="left"
+              className="px-3 py-1.5 text-xs"
+              sideOffset={10}
+            >
+              <span>{label}</span>
+            </TooltipContent>
+          </Tooltip>
+        </TooltipProvider>
+      )}
+    </>
+  );
+};
