+++ src/frontend/temp.env
@@ -1,0 +1,4 @@
+NEXTAUTH_SECRET=optivisionlab@hautoml
+NEXTAUTH_URL=""
+NEXT_PUBLIC_BASE_API=""
+NEXT_PUBLIC_TIMEOUT_CAROUSEL=15000
